pyproject.toml
[tool.poetry]
name = "img-to-base64"
version = "0.1.0"
description = ""
authors = ["alk-neq-me <toyko2001@gmail.com>"]
readme = "README.md"
packages = [{include = "img_to_base64"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
strawberry-graphql = {extras = ["debug-server"], version = "^0.142.2"}
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
